requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.1
web3==6.15.1
pillow==10.2.0
qrcode==7.4.2
opencv-python==4.9.0.80
reportlab==4.1.0
pikepdf==8.11.2
cloudinary==1.36.0
openpyxl==3.1.2
requests==2.31.0
pytz==2024.1
pdfplumber==0.10.3